moved old tests to proper test classes per command/feature
