+++ src/Tests/dotnet-new.Tests/DotnetNewInstantiate.cs
@@ -351,5 +351,166 @@ namespace Microsoft.DotNet.New.Tests
                 .And.NotHaveStdErr()
                 .And.HaveStdOutContaining("File actions would have been taken:");
         }
+
+        [Fact]
+        public void WhenSwitchIsSkippedThenItPrintsError()
+        {
+            Utils.CommandResult cmd = new DotnetNewCommand(Log)
+                .WithVirutalHive()
+                .Execute("Web1.1");
+
+            cmd.ExitCode.Should().NotBe(0);
+
+            if (!TestContext.IsLocalized())
+            {
+                cmd.StdErr.Should().StartWith("No templates found");
+            }
+        }
+
+        [Fact]
+        public void ItCanCreateTemplate()
+        {
+            string tempDir = TestUtils.CreateTemporaryFolder();
+            Utils.CommandResult cmd = new DotnetNewCommand(Log)
+                .WithVirutalHive()
+                .Execute("console", "-o", tempDir);
+            cmd.Should().Pass();
+        }
+
+        [Fact]
+        public void ItCanShowHelp()
+        {
+            Utils.CommandResult cmd = new DotnetNewCommand(Log)
+                .WithVirutalHive()
+                .Execute("--help");
+            cmd.Should().Pass()
+                .And.HaveStdOutContaining("Usage:")
+                .And.HaveStdOutContaining("dotnet new [command] [options]");
+        }
+
+        [Fact]
+        public void ItCanShowHelpForTemplate()
+        {
+            Utils.CommandResult cmd = new DotnetNewCommand(Log)
+                .WithVirutalHive()
+                .Execute("classlib", "--help");
+
+            cmd.Should().Pass()
+                .And.NotHaveStdOutContaining("Usage: new [options]")
+                .And.HaveStdOutContaining("Class Library (C#)")
+                .And.HaveStdOutContaining("--framework");
+        }
+
+        [Fact]
+        public void ItCanShowParseError()
+        {
+            Utils.CommandResult cmd = new DotnetNewCommand(Log)
+                .WithVirutalHive()
+                .Execute("update", "--bla");
+            cmd.Should().ExitWith(127)
+                .And.HaveStdErrContaining("Unrecognized command or argument '--bla'")
+                .And.HaveStdOutContaining("dotnet new update [options]");
+        }
+
+        [Fact]
+        public void WhenTemplateNameIsNotUniquelyMatchedThenItIndicatesProblemToUser()
+        {
+            Utils.CommandResult cmd = new DotnetNewCommand(Log)
+                .WithVirutalHive()
+                .Execute("c");
+
+            cmd.ExitCode.Should().NotBe(0);
+
+            if (!TestContext.IsLocalized())
+            {
+                cmd.StdErr.Should().StartWith("No templates found matching: 'c'.");
+            }
+        }
+
+        [Fact]
+        public void When_dotnet_new_is_invoked_multiple_times_it_should_fail()
+        {
+            string rootPath = TestUtils.CreateTemporaryFolder();
+
+            new DotnetNewCommand(Log)
+                .WithVirutalHive()
+                .WithWorkingDirectory(rootPath)
+                .Execute($"console", "--no-restore");
+
+            DateTime expectedState = Directory.GetLastWriteTime(rootPath);
+
+            CommandResult result = new DotnetNewCommand(Log)
+                .WithVirutalHive()
+                .WithWorkingDirectory(rootPath)
+                .Execute($"console", "--no-restore");
+
+            DateTime actualState = Directory.GetLastWriteTime(rootPath);
+
+            Assert.Equal(expectedState, actualState);
+
+            result.Should().Fail();
+        }
+
+        [Fact]
+        public void When_dotnet_new_is_invoked_with_preferred_lang_env_var_set()
+        {
+            string rootPath = TestUtils.CreateTemporaryFolder();
+
+            new DotnetNewCommand(Log)
+                .WithVirutalHive()
+                .WithWorkingDirectory(rootPath)
+                .WithEnvironmentVariable("DOTNET_NEW_PREFERRED_LANG", "F#")
+                .Execute($"console", "--no-restore", "-n", "f1")
+                .Should().Pass();
+
+            string expectedFsprojPath = Path.Combine(rootPath, "f1", "f1.fsproj");
+            Assert.True(File.Exists(expectedFsprojPath), $"expected '{expectedFsprojPath}' but was not found");
+        }
+
+        [Fact]
+        public void When_dotnet_new_is_invoked_default_is_csharp()
+        {
+            string rootPath = TestUtils.CreateTemporaryFolder();
+
+            new DotnetNewCommand(Log)
+                .WithVirutalHive()
+                .WithWorkingDirectory(rootPath)
+                .Execute($"console", "--no-restore", "-n", "c1")
+                .Should().Pass();
+
+            string expectedCsprojPath = Path.Combine(rootPath, "c1", "c1.csproj");
+            Assert.True(File.Exists(expectedCsprojPath), $"expected '{expectedCsprojPath}' but was not found");
+        }
+
+        [Fact]
+        public void Dotnet_new_can_be_invoked_with_lang_option()
+        {
+            string rootPath = TestUtils.CreateTemporaryFolder();
+
+            new DotnetNewCommand(Log)
+                .WithVirutalHive()
+                .WithWorkingDirectory(rootPath)
+                .Execute($"console", "--no-restore", "-n", "vb1", "-lang", "vb")
+                .Should().Pass();
+
+            string expectedCsprojPath = Path.Combine(rootPath, "vb1", "vb1.vbproj");
+            Assert.True(File.Exists(expectedCsprojPath), $"expected '{expectedCsprojPath}' but was not found");
+        }
+
+        [Fact]
+        public void When_dotnet_new_is_invoked_with_preferred_lang_env_var_empty()
+        {
+            string rootPath = TestUtils.CreateTemporaryFolder();
+
+            new DotnetNewCommand(Log)
+                .WithVirutalHive()
+                .WithWorkingDirectory(rootPath)
+                .WithEnvironmentVariable("DOTNET_NEW_PREFERRED_LANG", "")
+                .Execute($"console", "--no-restore", "-n", "c1")
+                .Should().Pass();
+
+            string expectedCsprojPath = Path.Combine(rootPath, "c1", "c1.csproj");
+            Assert.True(File.Exists(expectedCsprojPath), $"expected '{expectedCsprojPath}' but was not found");
+        }
     }
 }
